Let box bot inherit from parent
