+++ boxbot/Dockerfile
@@ -1,11 +1,11 @@
 FROM python:3.9
 
-RUN mkdir -p /usr/src/boxbot
-WORKDIR /usr/src/boxbot
+RUN mkdir -p /usr/src
+WORKDIR /usr/src
 
-COPY boxbot/requirements.txt /usr/src/boxbot
-RUN pip install --no-cache-dir -r requirements.txt
-
+COPY templates.py /usr/src/
 COPY boxbot /usr/src/boxbot
 
-ENTRYPOINT ["python", "boxbot.py"]
+RUN pip install --no-cache-dir -r boxbot/requirements.txt
+
+ENTRYPOINT ["python", "-m", "boxbot"]
